video to text/requirements.txt
# ===== Core dependencies =====
google-generativeai>=0.8.0
google-cloud-storage>=2.14.0
ffmpeg-python>=0.2.0
typer[all]>=0.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
aiofiles>=23.0.0

# ===== Web Dashboard =====
streamlit>=1.30.0
plotly>=5.18.0

# ===== YouTube 다운로드 =====
yt-dlp>=2024.1.0

# ===== Google Drive 다운로드 =====
gdown>=5.1.0

# ===== Development dependencies =====
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
ruff>=0.1.0
mypy>=1.8.0

# ===== RAG 시스템 =====
jamo>=0.4.1                        # 한글 자모 분리
google-api-python-client>=2.100.0  # Custom Search API

# ===== DOCX 출력 =====
python-docx>=1.0.0                 # Word 문서 생성

# ===== Phase 2: WhisperX (타임스탬프 정렬) =====
# GPU 필요 (CUDA). CPU도 가능하지만 느림
# pip install whisperx torch
# whisperx>=3.0.0
# torch>=2.0.0

# ===== Phase 2: LLM 교정 =====
# pip install anthropic
# anthropic>=0.18.0

# ===== Phase 2: Pyannote 화자분리 (WhisperX에 포함) =====
# Hugging Face 토큰 필요
# pip install pyannote.audio
# pyannote.audio>=3.1.0
